Rename StructInput
