+++ src/struct_input.rs
